Removed warnings, fixed timing units for Pynq
Should be the final version of pure SW unless there is some major upgrade possible. Just speeding up the multiplication puts a theoretical upper bound at about 30ms per complete key gen->encode->decode pipeline. This is surprisingly competitive for such a small thing to work on, and speaks to the placement of the multiplier as critical in the timing.